Добрался до кода наконец
